# docs/openapi.yml
openapi: 3.0.3
info:
  title: taskey API
  description: |-
    This is the API documentation for taskey.
    It will help you understand how to interact with the server.
    Source code for the project can be found at [taskey repository on Github](https://github.com/LassiHeikkila/taskey).
  contact:
    email: laheikki21@student.oulu.fi
  license:
    name: GNU General Public License
    url: https://raw.githubusercontent.com/LassiHeikkila/taskey/main/LICENSE
  version: 0.1.0
servers:
- url: http://localhost:8081/api/v1
security:
  - bearerAuth: []
paths:
...
  /{organization_id}/users/:
    get:
      tags:
      - users
      summary: Read users
      operationId: readUsers
      parameters:
      - $ref: '#/components/parameters/organizationId'
      responses:
        200:
          $ref: '#/components/responses/UsersResponse'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        501:
          $ref: '#/components/responses/Unimplemented'
    post:
      tags:
      - users
      summary: Create user
      operationId: createUser
      parameters:
      - $ref: '#/components/parameters/organizationId'
      requestBody:
        description: User object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        501:
          $ref: '#/components/responses/Unimplemented'
...
  /{organization_id}/users/{user_id}/tokens/:
    post:
      tags:
      - user tokens
      summary: Get a new token for user
      operationId: createUserToken
      parameters:
      - $ref: '#/components/parameters/organizationId'
      - name: user_id
        in: path
        description: id of the user
        required: true
        schema:
          type: string
          example: "user456"
      responses:
        200:
          $ref: '#/components/responses/UserTokenResponse'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        501:
          $ref: '#/components/responses/Unimplemented'
  /{organization_id}/users/{user_id}/tokens/{token}/:
    delete:
      tags:
      - user tokens
      summary: Delete a user token
      operationId: deleteUserToken
      parameters:
      - $ref: '#/components/parameters/organizationId'
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/token'
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        501:
          $ref: '#/components/responses/Unimplemented'
  /{organization_id}/machines/:
    get:
      tags:
      - machines
      summary: Read all machines
      operationId: readMachines
      parameters:
      - $ref: '#/components/parameters/organizationId'
      responses:
        200:
          $ref: '#/components/responses/MachinesResponse'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        501:
          $ref: '#/components/responses/Unimplemented'
    post:
      tags:
      - machines
      summary: Create a machine
      operationId: createMachineById
      parameters:
      - $ref: '#/components/parameters/organizationId'
      requestBody:
        description: Machine object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Machine'
        required: true
      responses:
        200:
          $ref: '#/components/responses/Success'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        501:
          $ref: '#/components/responses/Unimplemented'
...
  /{organization_id}/machines/{machine_id}/tokens/:
    post:
      tags:
      - machine tokens
      summary: Create a machine token
      operationId: createMachineToken
      parameters:
      - $ref: '#/components/parameters/organizationId'
      - $ref: '#/components/parameters/machineId'
      responses:
        200:
          $ref: '#/components/responses/MachineTokenResponse'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        501:
          $ref: '#/components/responses/Unimplemented'
  /{organization_id}/machines/{machine_id}/tokens/{token}/:
    delete:
      tags:
      - machine tokens
      summary: Revoke a machine token
      operationId: deleteMachineToken
      parameters:
      - $ref: '#/components/parameters/organizationId'
      - $ref: '#/components/parameters/machineId'
      - $ref: '#/components/parameters/token'
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        501:
          $ref: '#/components/responses/Unimplemented'
  /{organization_id}/machines/{machine_id}/schedule/:
    get:
      tags:
        - schedule
      summary: Read a machine's schedule
      operationId: readMachineSchedule
      parameters:
      - $ref: '#/components/parameters/organizationId'
      - $ref: '#/components/parameters/machineId'
      security:
      - bearerAuth: []
      - accessToken: []
      responses:
        200:
          $ref: '#/components/responses/ScheduleResponse'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        501:
          $ref: '#/components/responses/Unimplemented'
    post:
      tags:
        - schedule
      summary: Create a machine's schedule
      operationId: createMachineSchedule
      parameters:
      - $ref: '#/components/parameters/organizationId'
      - $ref: '#/components/parameters/machineId'
      requestBody:
        description: Schedule object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        501:
          $ref: '#/components/responses/Unimplemented'
    put:
      tags:
        - schedule
      summary: Update a machine's schedule
      operationId: updateMachineSchedule
      parameters:
      - $ref: '#/components/parameters/organizationId'
      - $ref: '#/components/parameters/machineId'
      requestBody:
        description: Schedule object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        501:
          $ref: '#/components/responses/Unimplemented'
    delete:
      tags:
        - schedule
      summary: Delete a machine's schedule
      operationId: deleteMachineSchedule
      parameters:
      - $ref: '#/components/parameters/organizationId'
      - $ref: '#/components/parameters/machineId'
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        501:
          $ref: '#/components/responses/Unimplemented'
  /{organization_id}/machines/self/schedule/:
      get:
        tags:
          - machine access
        summary: Endpoint for a machine to GET it's own schedule
        operationId: readMachineOwnSchedule
        parameters:
        - $ref: '#/components/parameters/organizationId'
        security:
        - accessToken: []
        responses:
          200:
            $ref: '#/components/responses/ScheduleResponse'
          401:
            $ref: '#/components/responses/Unauthenticated'
          403:
            $ref: '#/components/responses/Forbidden'
          404:
            $ref: '#/components/responses/NotFound'
          501:
            $ref: '#/components/responses/Unimplemented'
  /{organization_id}/tasks/:
    get:
      tags:
        - tasks
      summary: Read all defined tasks
      operationId: readAllTasks
      parameters:
      - $ref: '#/components/parameters/organizationId'
      security:
      - bearerAuth: []
      - accessToken: []
      responses:
        200:
          $ref: '#/components/responses/TasksResponse'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        501:
          $ref: '#/components/responses/Unimplemented'
    post:
      tags:
        - tasks
      summary: Create new task
      operationId: createTask
      parameters:
      - $ref: '#/components/parameters/organizationId'
      requestBody:
        description: Task object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
          $ref: '#/components/responses/Conflict'
        501:
          $ref: '#/components/responses/Unimplemented'
...
components:
  responses:
...
    Unauthenticated:
      description: Request could not be authenticated (invalid or no token in Authorization header)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            unauthenticated:
              $ref: '#/components/examples/UnauthenticatedRequest'
    Forbidden:
      description: Forbidden action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            unauthenticated:
              $ref: '#/components/examples/Forbidden'
...
    UsersResponse:
      description: array of user details
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              required:
              - payload
              properties:
                payload:
                  type: array
                  items:
                    $ref: '#/components/schemas/User'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessToken:
      type: apiKey
      name: Authorization
      in: header
  parameters:
    organizationId:
      name: organization_id
      in: path
      description: id of the organization
      required: true
      schema:
        type: string
        example: "org123"
...